Added device display reported specific refresh rate measurement

+++ common/platform/display.h
@@ -30,7 +30,7 @@ void display_set_ui_hidden(int hidden);
 
 int display_ui_is_hidden(void);
 
-int display_panel_refresh_hz(void);
+double display_panel_refresh_hz(void);
 
 int display_video_fast_path_allowed(void);
